Fixed emacs startup
Emacsclient won't call an alternate editor with command line
arguments, so using the -a switch didn't work.  However, it is easy to
tell if it exits with an error, so we do a quick calculation with it
to test if it's running and if it's not, start the daemon properly.