Fix errors caused by links that contain '%' characters
This error caused the linkify hook to exit prematurely, without
generating any output. The symptom was that the "*ERROR*: Not enough
arguments for format string" appeared on stderr and the rendered page
appeared empty.

Unfortunately, it is not possible to detect this kind of errors from
emacsclient status code. It is always zero.